requirements.txt
flask==3.0.3
python-docx==1.1.2
pypdf==5.1.0
pillow==10.4.0
pytesseract==0.3.13
